backend/requirements.txt
# Django & Django REST Framework
Django>=5.0.0,<5.1.0
djangorestframework>=3.14.0,<3.15.0
djangorestframework-simplejwt>=5.3.0,<5.4.0
django-cors-headers>=4.3.0,<4.4.0

# Database
psycopg2-binary>=2.9.9,<2.10.0

# Server
gunicorn>=21.2.0,<21.3.0
uvicorn>=0.24.0,<0.25.0

# Environment variables
python-dotenv>=1.0.0,<1.1.0

# Development & Testing
pytest>=7.4.0,<7.5.0
pytest-django>=4.5.2,<4.6.0
flake8>=6.1.0,<6.2.0
black>=23.11.0,<23.12.0
isort>=5.12.0,<5.13.0

# Utilities
Pillow>=10.1.0,<10.2.0  # For image processing
drf-yasg>=1.21.7,<1.22.0  # For API documentation